Implemented and tested a basix dataloader
