+++ tests/test_training.py
@@ -1,4 +1,5 @@
 import klax
+import jax.numpy as jnp
 import jax.random as jrandom
 import pytest
 
@@ -19,7 +20,7 @@ def test_dataloader(getkey):
 
     # Default batch size
     x = jrandom.uniform(getkey(), (33,))
-    data = (x, )
+    data = (x,)
     generator = klax.dataloader(data, key=getkey()) 
     assert next(generator)[0].shape[0] == 32
 
@@ -31,3 +32,19 @@ def test_dataloader(getkey):
     assert next(generator)[0].shape[0] == 2
     assert next(generator)[1][0].shape[0] == 10
     assert next(generator)[1][1].shape[0] == 2
+
+    # No batch dimensions
+    x = jrandom.uniform(getkey(), (10,))
+    data = (x,)
+    batch_mask = (False,)
+    with pytest.raises(ValueError):
+        generator = klax.dataloader(data, batch_mask=batch_mask, key=getkey()) 
+        next(generator)
+
+    # Different batch dimensions
+    x = jrandom.uniform(getkey(), (10,))
+    y = jrandom.uniform(getkey(), (5,))
+    data = (x, y)
+    with pytest.raises(ValueError):
+        generator = klax.dataloader(data, key=getkey()) 
+        next(generator)
